+++ test/dummy/app/models/order.rb
@@ -9,6 +9,7 @@ class Order < ApplicationRecord
     dimension :status
     dimension :region, through: :customer
     dimension :customer, through: :customer, column: :name
+    dimension :channel
     dimension :placed_on, granularity: :day
   end
 end